Add files via upload
This is the final version of my work on the Approximate solutions of the Schroedinger equation and Frobenius Method. This a project for the subject "Problems in Quantum Mechanics" in the Physics Department of the Aristotle University of Thessaloniki. For any comments or additional information do not hesitate to contact me at reviskostis@gmail.com.